+++ wxc_sdk/people/__init__.py
@@ -289,7 +289,7 @@ class PeopleApi(ApiChild, base='people'):
     """
 
     def list(self, email: str = None, display_name: str = None, id_list: list[str] = None, org_id: str = None,
-             roles: str = None, calling_data: bool = None, location_id: str = None,
+             roles: str = None, calling_data: bool = None, location_id: str = None, exclude_status: bool = None,
              **params) -> Generator[Person, None, None]:
         """
         List people in your organization. For most users, either the email or displayName parameter is required. Admin
@@ -327,6 +327,8 @@ class PeopleApi(ApiChild, base='people'):
         :type calling_data: bool
         :param location_id: List people present in this location.
         :type location_id: str
+        :param exclude_status: Omit people status/availability to enhance query performance.
+        :type exclude_status: bool
         :return: yield :class:`Person` instances
         """
         params.update((to_camel(k), v)
@@ -466,6 +468,10 @@ class PeopleApi(ApiChild, base='people'):
           in the response payload of `List People
           <https://developer.webex.com/docs/api/v1/people/list-people>`_ or `Get Person Details
           extension for a person.
+
+        * When updating a user with multiple email addresses using a PUT request, ensure that the primary email address
+          is listed first in the array. Note that the order of email addresses returned by a GET request is not
+          guaranteed..
 
         * The People API is a combination of several microservices, each responsible for specific attributes of a
           person. As a result, a PUT request that returns an error response code may still have altered some values of